requirements.txt
Django==5.0.6
Pillow==10.3.0
Requests==2.32.3
ytmusicapi==1.7.3
python-decouple==3.8